requirements.txt
gspread==6.0.0
oauth2client==4.1.3
openai>=1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0 